Fixed creation with known identifier.
